refactor(dialogue): split turn and summary hooks
